+++ package.json
@@ -32,12 +32,14 @@
     "react-hook-form": "^7.47.0",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^4.11.0",
+    "react-infinite-scroll-component": "^6.1.0",
+    "react-loader-spinner": "^6.1.6",
     "react-lottie": "^1.2.4",
     "react-otp-input": "^3.1.0",
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.16.0",
     "react-scrollable-feed": "^1.3.2",
-    "socket.io-client": "^4.7.2",
+    "socket.io-client": "^4.2.0",
     "tailwind-merge": "^2.3.0",
     "tailwindcss": "^3.3.3"
   },
