+++ server/controllers/friend.js
@@ -239,41 +239,3 @@ exports.getFriends = async (req, res) => {
         });
     }
 }
-
-exports.getFriendsPosts = async (req, res) => {
-    try {
-        const { id } = req.user
-        if (!id) {
-            throw new Error('id not found')
-        }
-        const user = await User.findById(id).select('friends').exec();
-        const friendIds = user.friends
-
-        if (!friendIds.length) {
-            return res.status(200).json({
-                success: true,
-                message: 'No Friends Found',
-                data: []
-            });
-        }
-
-        const posts = await Post
-            .find({ user: { $in: friendIds } })
-            .populate('likes')
-            .populate({ path: 'user', populate: 'profileDetails' })
-            .exec();
-
-        const shuffledPosts = shuffleArray(posts)
-
-        res.status(200).json({
-            success: true,
-            message: 'posts fetched successfully',
-            data: shuffledPosts
-        });
-    } catch (error) {
-        res.status(500).json({
-            success: false,
-            message: error.message
-        })
-    }
-}
